requirements.txt
asgiref==3.3.4
cloudinary==1.25.0
dj-rest-auth==2.1.9
Django==3.2.4
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
oauthlib==3.1.1
Pillow==8.2.0
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1